+++ packages/integration-tests/Granules.js
@@ -12,12 +12,31 @@ const granulesApi = require('@cumulus/api-client/granules');
 const pick = require('lodash/pick');
 const pRetry = require('p-retry');
 
+class GranuleNotFoundError extends Error {
+  constructor(id) {
+    super(`Granule ${id} not found`);
+    Error.captureStackTrace(this, this.constructor);
+    this.name = this.constructor.name;
+  }
+}
+
+const getGranule = async (params) => {
+  const response = await granulesApi.getGranule(
+    pick(params, ['prefix', 'granuleId', 'callback'])
+  );
+
+  if (response.status === 404) throw new GranuleNotFoundError(params.granuleId);
+
+  return JSON.parse(response.body);
+};
+
 /**
- * Wait for a granule's status to be `completed` and return the granule
+ * Wait for a granule to have an expected status and return the granule
  *
  * @param {Object} params
  * @param {string} params.prefix - the name of the Cumulus stack
  * @param {string} params.granuleId - the `granuleId` of the granule
+ * @param {string} params.status - the status to wait for
  * @param {Function} [params.callback=cumulusApiClient.invokeApifunction] - an async function to
  * invoke the API Lambda that takes a prefix / user payload
  * @param {integer} [params.timeout=30] - the number of seconds to wait for the
@@ -26,32 +45,24 @@ const pRetry = require('p-retry');
  *
  * @alias module:Granules
  */
-const getCompletedGranule = async (params = {}) =>
+const getGranuleWithStatus = async (params = {}) =>
   pRetry(
     async () => {
       let granule;
 
       try {
-        const response = await granulesApi.getGranule(
-          pick(params, ['prefix', 'granuleId', 'callback'])
-        );
-
-        granule = JSON.parse(response.body);
+        granule = await getGranule(pick(params, ['prefix', 'granuleId', 'callback']));
       } catch (err) {
         throw new pRetry.AbortError(err);
       }
 
-      if (granule.status === 'completed') return granule;
+      if (granule.status === params.status) return granule;
 
-      if (granule.status === 'failed') {
+      if (['completed', 'failed'].includes(granule.status)) {
         throw new pRetry.AbortError(
-          new Error(`Granule ${params.granuleId} failed`)
-        );
-      }
-
-      if (granule.statusCode === 404) {
-        throw new pRetry.AbortError(
-          new Error(`Granule ${params.granuleId} not found`)
+          new Error(
+            `Expected granule ${params.granuleId} to have status ${params.status} but found ${granule.status}`
+          )
         );
       }
 
@@ -64,5 +75,5 @@ const getCompletedGranule = async (params = {}) =>
   );
 
 module.exports = {
-  getCompletedGranule
+  getGranuleWithStatus
 };